feat : CoinService 의 Coin 수정, 삭제 권한 체크 구현

+++ cotobang/src/main/java/com/example/cotobang/service/CoinService.java
@@ -4,6 +4,7 @@ import com.example.cotobang.domain.Coin;
 import com.example.cotobang.domain.User;
 import com.example.cotobang.dto.CoinDto;
 import com.example.cotobang.errors.CoinNotFoundException;
+import com.example.cotobang.errors.NotAuthorityException;
 import com.example.cotobang.respository.CoinRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -52,7 +53,9 @@ public class CoinService {
     public Coin updateCoin(Long targetId, CoinDto source, User user) {
         Coin coin = getCoin(targetId);
 
-        // 유저 비교 구현
+        if (!user.compare(coin.getUser())) {
+            throw new NotAuthorityException(user.getName());
+        }
 
         coin.change(
                 source.getMarket(),
@@ -67,7 +70,9 @@ public class CoinService {
     public Coin deleteCoin(Long id, User user) {
         Coin coin = getCoin(id);
 
-        // 유저 비교 구현
+        if (!user.compare(coin.getUser())) {
+            throw new NotAuthorityException(user.getName());
+        }
 
         coinRepository.delete(coin);
